OpenEye/backend/requirements.txt
fastapi
google-generativeai
pillow
python-multipart
transformers
uvicorn
